% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darmanis_Brain_scRNAseq_Annotations.R
\docType{data}
\name{Darmanis_Brain_scRNAseq_Annotations}
\alias{Darmanis_Brain_scRNAseq_Annotations}
\title{Human Brain single cell RNA-Seq data from Darmanis et. al. (2015)}
\format{A data frame with 461 rows and 6 variables}
\usage{
data(Darmanis_Brain_scRNAseq_Annotations)
}
\description{
Annotations describing the brain samples and experimental settings.
}
\keyword{datasets}
