% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darmanis_Brain_scRNAseq_Read_Counts.R
\docType{data}
\name{Darmanis_Brain_scRNAseq_Read_Counts}
\alias{Darmanis_Brain_scRNAseq_Read_Counts}
\title{Human Brain single cell RNA-Seq data from Darmanis et. al. (2015)}
\format{A data frame with 23326 rows and 461 variables}
\usage{
data(Darmanis_Brain_scRNAseq_Read_Counts)
}
\description{
Gene-level read counts from the single cell RNA sequencing experiment of Darmanis et. al. (2015)
}
\keyword{datasets}
