% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lawlor_Islet_scRNAseq_Annotations.R
\docType{data}
\name{Lawlor_Islet_scRNAseq_Annotations}
\alias{Lawlor_Islet_scRNAseq_Annotations}
\title{Human islet single cell RNA-Seq data from Lawlor et. al. (2017)}
\format{A data frame with 638 rows and 26 variables}
\usage{
data(Lawlor_Islet_scRNAseq_Annotations)
}
\description{
Annotations describing the islet samples and experimental settings.
}
\keyword{datasets}
