% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lawlor_Islet_scRNAseq_Read_Counts.R
\docType{data}
\name{Lawlor_Islet_scRNAseq_Read_Counts}
\alias{Lawlor_Islet_scRNAseq_Read_Counts}
\title{Human islet single cell RNA-Seq data from Lawlor et. al. (2017)}
\format{A data frame with 26616 rows and 638 variables}
\usage{
data(Lawlor_Islet_scRNAseq_Read_Counts)
}
\description{
Gene-level read counts from the single cell RNA sequencing experiment of Lawlor et. al. (2017)
}
\keyword{datasets}
