% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patel_Glioblastoma_scRNAseq_Annotations.R
\docType{data}
\name{Patel_Glioblastoma_scRNAseq_Annotations}
\alias{Patel_Glioblastoma_scRNAseq_Annotations}
\title{Human Glioblastoma Single Cell RNA-Seq Data from Patel et. al. (2014)}
\format{A data frame with 434 rows and 3 variables}
\usage{
data(Patel_Glioblastoma_scRNAseq_Annotations)
}
\description{
Annotations describing the samples and experimental settings.
}
\keyword{datasets}
