% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patel_Glioblastoma_scRNAseq_Read_Counts.R
\docType{data}
\name{Patel_Glioblastoma_scRNAseq_Read_Counts}
\alias{Patel_Glioblastoma_scRNAseq_Read_Counts}
\title{Human Glioblastoma Single Cell RNA-Seq Data from Patel et. al. (2014)}
\format{A data frame with 25415 rows and 434 variables}
\usage{
data(Patel_Glioblastoma_scRNAseq_Read_Counts)
}
\description{
Gene-level read counts from the single cell RNA sequencing experiment of Patel et. al. (2014)
}
\keyword{datasets}
