% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xin_Islet_scRNAseq_Annotations.R
\docType{data}
\name{Xin_Islet_scRNAseq_Annotations}
\alias{Xin_Islet_scRNAseq_Annotations}
\title{Human islet single cell RNA-Seq data from Xin et. al. (2016)}
\format{A data frame with 1600 rows and 9 variables}
\usage{
data(Xin_Islet_scRNAseq_Annotations)
}
\description{
Annotations describing the islet samples and experimental settings.
}
\keyword{datasets}
