% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xin_Islet_scRNAseq_Read_Counts.R
\docType{data}
\name{Xin_Islet_scRNAseq_Read_Counts}
\alias{Xin_Islet_scRNAseq_Read_Counts}
\title{Human islet single cell RNA-Seq data from Xin et. al. (2016)}
\format{A data frame with 26616 rows and 1600 variables}
\usage{
data(Xin_Islet_scRNAseq_Read_Counts)
}
\description{
Gene-level read counts from the single cell RNA sequencing experiment of Xin et. al. (2016)
}
\keyword{datasets}
